@@ -5,8 +5,8 @@ import 'dart:math';
 /// returns map of smoothed points
 Map<double, double> custom(
   Map<double, double> input,
-  int width,
-  Function(double distance) weighting,
+  int? width,
+  Function(double distance)? weighting,
 ) {
   Map<double, double> smoothed = Map();
 
@@ -15,7 +15,7 @@ Map<double, double> custom(
   for (int i = 0; i < input.entries.length; i++) {
     final MapEntry<double, double> point = entries[i];
     final Iterable<MapEntry<double, double>> neighbours = entries.getRange(
-      max(0, i - width),
+      max(0, i - width!),
       min(i + width, input.length - 1),
     );
     final int localCenter = min(i, width);
@@ -26,7 +26,7 @@ Map<double, double> custom(
     for (int j = 0; j < neighbours.length; j++) {
       final int distance = (localCenter - j).abs();
       final double percent = distance / width;
-      final double weight = weighting(percent);
+      final double weight = weighting!(percent);
 
       if (weight <= 0.01) {
         if (i < j) break;
